requirements.txt
discord-py-interactions[speedup]==5.8.0
tansy==0.7.0
python-dotenv==1.0.0
humanize==4.6.0
Pillow==9.5.0
uvloop==0.17.0; platform_system == "Linux"